Fix maximum recursion error (fixes #80)
In 0da4f64b2, JSON validation was introduced. In order to acheive this a
validate_payload() decorator was getting added to every Resource(). The
problem was because Resource() was getting instantiated new every time
it would just continue to add this method over and over until eventually
Python would die due to a maximum recursion error.

By moving the validate_payload() method to the Api() class it resolves
the issue of a new version of the method being added to each Resource().
In order to keep the ability of being able to easily change/override the
validate_payload() method a _add_validate_payload() method was added to
the Resource() init method where an end user can easily manipulate the
workflow in any subclasses.