Add semantic validation for an input payload
The JSON Schema specification defines the optionally-supported 'format'
keyword for semantic validation of string values. [1]

For example, given the following JSON schema:
{
  "type": "object",
  "properties": {
    "name": {"type": "string", "format": "ipv4"}
  }
}
A JSON schema validator that does not conform to the 'format' keyword would
render the JSON object {"ip": "awesome-ip"} valid, whereas one that
conforms to it, would render it invalid.

Currently Flask-RESTPlus does not support the ability to add a semantic
validator. With this patch, the Api constructor may receive a
`jsonschema.FormatChecker` object as argument, that will be passed from
the `resource.Resource.validate_payload` method to the
`model.Model.validate` method, and will be used in the invocation of the
Draft4Validator that actually performs the validation.

[1] http://tools.ietf.org/html/draft-fge-json-schema-validation-00#section-7