Add semantic validation for an input payload (#178)
Currently Flask-RESTPlus does not support the ability to add a semantic
validator. With this patch, the Api constructor may receive a
`jsonschema.FormatChecker` object as argument, that will be passed from
the `resource.Resource.validate_payload` method to the
`model.Model.validate` method, and will be used in the invocation of the
Draft4Validator that actually performs the validation.