Fix $ref issue of not importing Json SchemaModels
There are no more errors shown in Swagger Doc and the models sections in Swagger Doc show correctly the sub-schema in $ref fields. You will still need to register all models and sub-models that you expect on response or request. You register them with: api.schema_model(<NameOfModel>, <JSONSchema>)
This fix fixes the errors found in the the example found in the restplus documentation:
http://flask-restplus.readthedocs.io/en/stable/marshalling.html#define-model-using-json-schema

The fix related to the issue: JSON Schema $ref not working (not even the doc example) #275