Fix late resource (route) registration
When `Api` is initialized without `app` and `init_app` is called before all the resources are registered, the resources end up to be 404 Not Found.

Consider the following code:

```python
api = Api(...)

app = Flask('app')
api.init_app(app)

ns = Namespace(...)
ns.route(...)

api.add_namespace(ns)
```

Without this fix, the `ns` routes end up to be not propagated to the Flask `app`, and thus the server responds with 404 Not Found.