Modifies Api.error_router so that exceptions raised in error handlers are not suppressed in the the logs.
Prior to this change, exceptions raises in error handlers registered to an Api were suppressed, and the
original exception which caused the error handler to be invoked was reported in logs. This made it
very challenging to diagnose programming errors in error handlers, or even identify that an error handler
implementation was faulty in the first place.

With this change, programming mistakes and other exceptions raised in error handlers are properly reported,
and a 500 Internal Server Error returned. The exception which was being handled is chained to this subsequent
error and also reported.