Make test_error_router_falls_back_to_original that exceptions in errorhandlers registers on Api are not suppressed.
Prior to this change, this test did indeed check that the error router fell back to the original handler, but
it also (incorrectly, in my view) asserted that the exception passed to that handler was the error that caused
the handler to be invoked. This suppressed any errors in error handler implementation, making them hard to get
right. Now we assert that any programming (or other) exceptions raises in the error handler are reported. The
original hander-invoking error is chained to this error, so there is no loss of information.